+++ source/CHANGELOG.md
@@ -1,4 +1,13 @@
+## 0.0.39  
+some changes on the api  
+changes: 
+- `everyWeekDayStartingFromNMonthDay` replaced for `everyNDays`  
+- `everyWeekDay` chage behaviour, now does not accept parameters, it jsut set `dow` to `2-6`
 
+additions
+- `everyWeekDay` to set `dow` to `2-6`
+- `atWeekDay` and `atWeekDayAdd`
+- `betweenWeekDays`
 
 ## 0.0.38  
 all 7 range resolvers are exposed in static `Quartzcron.solvers`